requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.5.0
html2text==2020.1.16
idna==2.10
isort==5.6.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.0
psutil==5.7.3
py-cpuinfo==7.0.0
pylint==2.6.0
python-dotenv==0.15.0
regex==2020.10.11
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
yarl==1.5.1
